# docs/swagger.yaml
definitions:
  entity.Comment:
    properties:
      content:
        type: string
      created_at:
        type: integer
      id:
        type: string
      post_id:
        type: integer
      updated_at:
        type: integer
      user_id:
        type: integer
    type: object
  handler.UploadImageResponse:
    properties:
      path:
        type: string
      url:
        type: string
    type: object
...
  request.CreateCommentRequest:
    properties:
      content:
        type: string
      post_id:
        type: integer
    required:
    - content
    - post_id
    type: object
  request.CreateFavoriteRequest:
    properties:
      post_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  request.UpdateCommentRequest:
    properties:
      content:
        type: string
      post_id:
        type: integer
    required:
    - content
    - post_id
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/blogs/upload/image:
    post:
      consumes:
      - multipart/form-data
      description: 'Upload an image file for a blog post (admin only). Maximum file
        size: 5MB. Supported formats: JPEG, PNG, GIF, WebP'
      parameters:
      - description: Image file to upload
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Image uploaded successfully
          schema:
            $ref: '#/definitions/handler.UploadImageResponse'
        "400":
          description: Invalid file or file too large
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to upload file
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Upload blog image
      tags:
      - Blogs
...
  /api/v1/comments:
    post:
      consumes:
      - application/json
      description: Add a comment to a blog post
      parameters:
      - description: Comment creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Comment created successfully
          schema:
            $ref: '#/definitions/entity.Comment'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a new comment
      tags:
      - Comments
  /api/v1/comments/{id}:
    delete:
      description: Remove a comment from a blog post
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comment deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid comment ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete a comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update an existing comment on a blog post
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Comment updated successfully
          schema:
            $ref: '#/definitions/entity.Comment'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update a comment
      tags:
      - Comments
  /api/v1/favorite:
    post:
      consumes:
      - application/json
      description: Add a blog to the user's favorites
      parameters:
      - description: Favorite creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateFavoriteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Favorite created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a new favorite
      tags:
      - Favorites
  /api/v1/favorite/{favoriteId}:
    delete:
      description: Remove a blog from the user's favorites
      parameters:
      - description: Favorite ID
        in: path
        name: favoriteId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Favorite deleted successfully
        "400":
          description: Invalid favorite ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Favorite not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete a favorite
      tags:
      - Favorites
  /api/v1/favorite/{userId}:
    get:
      description: Retrieve all favorite blogs for a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of user's favorite blogs
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get user's favorites
      tags:
      - Favorites
...
  /images/{filename}:
    get:
      description: Retrieve a blog image by filename (public endpoint)
      parameters:
      - description: Image filename
        in: path
        name: filename
        required: true
        type: string
      produces:
      - image/jpeg
      - image/png
      - image/gif
      - image/webp
      responses:
        "200":
          description: Image file
        "400":
          description: Invalid filename
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Image not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve image
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get blog image
      tags:
      - Blogs
swagger: "2.0"