# modules/favorites/handler/favorite_handler.go
package handler

import (
	"go-modular/internal/pkg/bus"
	"go-modular/internal/pkg/logger"
	"go-modular/internal/pkg/middleware"
	"go-modular/internal/pkg/utils"
	"go-modular/modules/favorites/domain/service"

	"github.com/labstack/echo/v4"
)

type FavoriteHandler struct {
	favoriteService service.FavoriteService
	log             *logger.Logger
	event           *bus.EventBus
	r               *utils.Response
}

func NewFavoriteHandler(log *logger.Logger, event *bus.EventBus, favoriteService service.FavoriteService) *FavoriteHandler {
	return &FavoriteHandler{
		favoriteService: favoriteService,
		log:             log,
		event:           event,
		r:               nil,
	}
}

// CreateFavorite creates a new favorite
// @Summary Create a new favorite
// @Description Add a blog to the user's favorites
// @Tags Favorites
// @Accept json
// @Produce json
// @Param request body request.CreateFavoriteRequest true "Favorite creation request"
// @Success 201 {object} map[string]interface{} "Favorite created successfully"
// @Failure 400 {object} map[string]interface{} "Invalid request body"
// @Failure 401 {object} map[string]interface{} "Unauthorized"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /api/v1/favorite [post]
// @Security Bearer
func (h *FavoriteHandler) CreateFavorite(c echo.Context) error {
	return nil
}

// DeleteFavorite deletes a favorite
// @Summary Delete a favorite
// @Description Remove a blog from the user's favorites
// @Tags Favorites
// @Produce json
// @Param favoriteId path int true "Favorite ID"
// @Success 204 "Favorite deleted successfully"
// @Failure 400 {object} map[string]interface{} "Invalid favorite ID"
// @Failure 401 {object} map[string]interface{} "Unauthorized"
// @Failure 404 {object} map[string]interface{} "Favorite not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /api/v1/favorite/{favoriteId} [delete]
// @Security Bearer
func (h *FavoriteHandler) DeleteFavorite(c echo.Context) error {
	return nil
}

// GetByUserId gets all favorites for a user
// @Summary Get user's favorites
// @Description Retrieve all favorite blogs for a specific user
// @Tags Favorites
// @Produce json
// @Param userId path int true "User ID"
// @Success 200 {array} map[string]interface{} "List of user's favorite blogs"
// @Failure 400 {object} map[string]interface{} "Invalid user ID"
// @Failure 401 {object} map[string]interface{} "Unauthorized"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /api/v1/favorite/{userId} [get]
// @Security Bearer
func (h *FavoriteHandler) GetByUserId(c echo.Context) error {
	return nil
}

func (h *FavoriteHandler) RegisterRoutes(e *echo.Echo, basePath string) {
	favorite := e.Group(basePath+"/favorite", middleware.Auth)
	favorite.GET("/:userId", h.GetByUserId)
	favorite.POST("", h.CreateFavorite)
	favorite.DELETE("/:favoriteId", h.DeleteFavorite)
}